requirements.txt
python-dotenv==0.19.2
openai==2.24.0
requests==2.31.0
questionary==1.10.0
PyYAML==6.0.2